Create LoginJoin
It is a combined file(MemberLogin+LoginForm). created two Frames
(which have ImagePanel each) so that first one has a setVisible(true)
as a default and later got setVisible(false) which makes the other one 
setVisible(true). but I can not see anything but first gray frame...
and somewhat it tries to connect itself to server which calls 'connect
server failure which is for sure without server on.
failure